Add MySQL and es link and change controller DS
I changed the ArticleController methods' data structure to be strings
so as to be able to keep the order that is found in the es documents
because it is crucial to the article restitution in the frontend.
I also exposed an endpoint in the REST api that handles the linking
of the MySQL table with the elasticsearch index.

DISCLAIMER : The linking endpoint should only be used once because it
only links the 974 articles that are in the es index to the MySQL table.