Link login system to frontend
I linked the login system to the frontend using JWT while also
saving the user session in the local storage. I also handled CORS
on the backend side to allow the frontend endpoint as an origin for
all the backend endpoints. In addition to this, I also added a logout
functionnality that enables the user to logout by deleting their JWT
token from the brower's local storage.